backend/requirements.txt
# Core AI và Machine Learning
dlib==19.24.2
insightface==0.7.3
tensorflow==2.14.0
opencv-contrib-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
onnxruntime-gpu==1.16.3

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.1.2

# Data Validation & Utils
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3

# Image Processing
Pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

